requirements/base.txt
Flask==0.10.1
Flask-HTTPAuth==2.2.0
Flask-RESTful==0.2.5
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
amqp==1.3.3
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.8
celery==3.1.5
docker-py==0.3.0
gunicorn==18.0
itsdangerous==0.23
kombu==3.0.6
pytz==2013.8
redis==2.8.0
requests==2.2.1
six==1.5.2
websocket-client==0.11.0
wsgiref==0.1.2
